Extracted function
Moved the video_file_condition from find_media_files_in_dir.py to its own script file--file_condition_functions.py.
