+++ PlexFileOrganizer/functions/__init__.py
@@ -7,4 +7,5 @@ from .custom_objects import MediaFile, ExtraFolders
 from .directory_scanner import directory_scanner
 from .correct_media_file_format import FolderAndFilePatterns
 from .generate_correct_video_file_format import generate_correct_video_file_format
-from .find_media_files_in_dir import find_media_files_in_dir, video_file_condition
+from .find_media_files_in_dir import find_media_files_in_dir
+from .file_condition_functions import video_file_condition
