+++ PlexFileOrganizer/functions/find_media_files_in_dir.py
@@ -1,5 +1,6 @@
 import pathlib
-from PlexFileOrganizer.functions import directory_scanner
+from PlexFileOrganizer.functions import directory_scanner, video_file_condition
+
 
 def find_media_files_in_dir(file_condition, directory):
     """
@@ -30,19 +31,6 @@ def find_media_files_in_dir(file_condition, directory):
 
     # this is necessary for when scanning a folder with just files or reaching the final folder of the directory
     yield tuple(files_in_directory)
-
-# *** File Condition Functions ***
-def video_file_condition(file_path):
-    """
-    The condition that validates that the file found is a video file.
-
-    :param file_path: The directory location of the file
-    :return:
-    """
-    if any(file_path.endswith(file_extension) for file_extension in ['.mkv', '.mp4', '.avi']):
-        return True
-    else:
-        return False
 
 if __name__ == '__main__':
     dir_path = "/Volumes/Media Library/Anime TV Shows/Frieren Beyond Journey's End"
